Rename connection module to conn
