+++ src/req/nblock.rs
@@ -6,7 +6,7 @@ use serde::Serialize;
 use crate::memo::TcpWriteBuff;
 use crate::service::{MsgPackCodec, ReplService};
 
-use crate::connection::Connection;
+use crate::conn::Connection;
 use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
 use log::{error, trace};
 use mio::net::{TcpListener, TcpStream};
